Next.js `16.x` upgrade
