+++ ws-nextjs-app/.babelrc
@@ -13,6 +13,7 @@
     "@emotion/babel-plugin",
     "@babel/plugin-proposal-export-default-from",
     "@babel/plugin-transform-private-methods",
+    "@babel/plugin-transform-class-static-block",
     [
       "transform-rename-import",
       { "original": "@loadable/component", "replacement": "next/dynamic" }
