+++ ws-nextjs-app/next.config.js
@@ -47,9 +47,6 @@ module.exports = {
     LOG_TO_CONSOLE: 'true',
     NEXTJS: 'true',
   },
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
   /*
    Requires pages that are routed to have the .page extension, e.g. [variant].page.tsx,
    which allows for co-locating components within the pages directory, e.g. styles.ts
